+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.4
+
+- Fix the Curl.tryParse method
+- Set ArgParser.allowTrailingOptions to true
+- Fix bug when the command line arguments splitter can't handle the last double quotes
+
 ## 1.0.3
 
 - Support \" inside of the "